requirements.txt
streamlit
pillow
ultralytics
opencv-python
google-generativeai
python-dotenv